requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
sqlparse==0.5.3
tzdata==2025.2
